We added a new repository to the collection of programing resources.
We started doing a Windows Form with a TextBox with a black background simuling a CMD.

It has some components like a label with a double click event, a numeric up down, a listbox what says when he get/loss the focus,
and finally a picture box with an Enter / Leaves event.

We saw some events in that practice, every time an event triggers it shows a message on our TextBox "Console".